All tested and blacked. Maybe change imshow to pcolormesh.
