+++ runs/minimal_example.py
@@ -21,33 +21,28 @@ srf_flx = ideal_source(nxy, domain)
 
 z, profs = vertical_profiles(nz, meas_height, wind, ustar)
 
-srf_conc, bg_conc, conc, flx = steady_state_transport_solver(
-    srf_flx, z, profs, domain, nz
-)
+grid, conc, flx = steady_state_transport_solver(srf_flx, z, profs, domain, nz)
+
+Z, Y, X = grid
 
 if __name__ == "__main__":
     logger.info("Minimal example for neutrally stratified BL and default settings.")
     logger.info("")
-    plt.figure()
-    plt.imshow(srf_conc, origin="lower", extent=[0, domain[0], 0, domain[1]])
-    plt.title("Concentration at z0")
-    plt.xlabel("x")
-    plt.ylabel("y")
-    plt.colorbar()
-    plt.savefig("plots/concentration_at_z0.png")
 
     plt.figure()
-    plt.imshow(conc, origin="lower", extent=[0, domain[0], 0, domain[1]])
-    plt.title("Concentration at zm")
-    plt.xlabel("x")
-    plt.ylabel("y")
+    plt.pcolormesh(X, Y, conc)
+    plt.title(f"Concentration at {meas_height} m")
+    plt.xlabel("x [m]")
+    plt.ylabel("y [m]")
+    plt.gca().set_aspect("equal")
     plt.colorbar()
-    plt.savefig("plots/concentration_at_zm.png")
+    plt.savefig("plots/concentration_at_meas_height.png")
 
     plt.figure()
-    plt.imshow(flx, origin="lower", extent=[0, domain[0], 0, domain[1]])
-    plt.title("Vertical kinematic flux at zm")
-    plt.xlabel("x")
-    plt.ylabel("y")
+    plt.pcolormesh(X, Y, flx)
+    plt.title(f"Vertical kinematic flux at {meas_height} m")
+    plt.xlabel("x [m]")
+    plt.ylabel("y [m]")
+    plt.gca().set_aspect("equal")
     plt.colorbar()
-    plt.savefig("plots/kinematic_flux_at_zm.png")
+    plt.savefig("plots/kinematic_flux_at_meas_height.png")
