+++ src/bldfm/__init__.py
@@ -1,7 +1,6 @@
 from pathlib import Path
 
 from .solver import steady_state_transport_solver
-from .solver3d import steady_state_transport_solver_3d
 from .utils import (
     compute_wind_fields,
     point_source,
@@ -17,7 +16,6 @@ setup_logging()
 
 __all__ = [
     "steady_state_transport_solver",
-    "steady_state_transport_solver_3d",
     "compute_wind_fields",
     "point_source",
     "ideal_source",
